# Removing Vue.js (for now, anyway)#
As there were far too many issues getting vue.js to work with the Features & Options list on the Product page, I have chosen to remove it for this project. There may be a possibility for adding it back in the future, but it is not even needed right now, given Shawn's requirements for the Product page.

* Adds `Prettier` and runs it on a couple of files.

Signed-off-by: Chris McGee <chris@chrismcgee.info>